requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
decorator==5.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0     
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
drf-yasg==1.21.4
dynaconf==3.1.11
flake8==5.0.4
hvac==1.0.2
idna==3.4
importlib-resources==5.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
mccabe==0.7.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
psycopg2==2.9.4
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyhcl==0.4.4
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
shortuuid==1.0.9
simplejson==3.17.6
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
win32-setctime==1.1.0
zipp==3.10.0